Move batch resources to reduce pipeline
