+++ src/topology/generic_pipeline.rs
@@ -7,11 +7,10 @@ use crate::topology::flush_control::{FlushReceiver, conditional_flush};
 use crate::topology::payload::Message;
 use opentelemetry_proto::tonic::common::v1::any_value::Value::StringValue;
 use opentelemetry_proto::tonic::common::v1::{AnyValue, KeyValue};
-use opentelemetry_proto::tonic::logs::v1::{ResourceLogs, ScopeLogs};
-use opentelemetry_proto::tonic::metrics::v1::metric::Data;
-use opentelemetry_proto::tonic::metrics::v1::{ResourceMetrics, ScopeMetrics};
+use opentelemetry_proto::tonic::logs::v1::ResourceLogs;
+use opentelemetry_proto::tonic::metrics::v1::ResourceMetrics;
 use opentelemetry_proto::tonic::resource::v1::Resource;
-use opentelemetry_proto::tonic::trace::v1::{ResourceSpans, ScopeSpans};
+use opentelemetry_proto::tonic::trace::v1::ResourceSpans;
 #[cfg(feature = "pyo3")]
 use rotel_sdk::model::{PythonProcessable, register_processor};
 #[cfg(feature = "pyo3")]
@@ -265,7 +264,7 @@ where
                             .collect();
                         debug!(batch_size = to_send.len(), "Flushing a batch in timout handler");
 
-                        let fut = self.fanout.async_send(to_send);
+                        let fut = self.fanout.send_async(to_send);
                         send_fut = Some(fut);
                     }
                 },
@@ -286,7 +285,7 @@ where
                             .collect();
 
                         // We are exiting, so we just want to log the failure to send
-                        if let Err(e) = self.fanout.async_send(remain_batch).await {
+                        if let Err(e) = self.fanout.send_async(remain_batch).await {
                             error!(error = ?e, "Unable to send item while exiting, will drop data.");
                         }
 
@@ -337,7 +336,7 @@ where
                             let to_send: Vec<T> = popped.into_iter()
                                 .flat_map(|msg| msg.payload)
                                 .collect();
-                            let fut = self.fanout.async_send(to_send);
+                            let fut = self.fanout.send_async(to_send);
                             send_fut = Some(fut);
                         }
                         Ok(None) => {},
@@ -371,7 +370,7 @@ where
                                     .collect();
                                 debug!(batch_size = to_send.len(), "Flushing a batch on flush message");
 
-                                if let Err(e) = self.fanout.async_send(to_send).await {
+                                if let Err(e) = self.fanout.send_async(to_send).await {
                                     error!(error = ?e, "Unable to send item, exiting.");
                                     return Err(format!("Pipeline was unable to send downstream: {}", e).into())
                                 }
@@ -404,175 +403,6 @@ where
     match send_fut {
         None => None,
         Some(fut) => Some(fut.await),
-    }
-}
-
-impl BatchSizer for ResourceSpans {
-    fn size_of(&self) -> usize {
-        self.scope_spans.iter().map(|sc| sc.spans.len()).sum()
-    }
-}
-
-impl BatchSizer for [ResourceSpans] {
-    fn size_of(&self) -> usize {
-        self.iter().map(|n| n.size_of()).sum()
-    }
-}
-
-impl BatchSplittable for ResourceSpans {
-    fn split(&mut self, split_n: usize) -> Self
-    where
-        Self: Sized,
-    {
-        let mut count_moved = 0;
-        let mut split_scope_spans = vec![];
-        while !self.scope_spans.is_empty() && count_moved < split_n {
-            // first just check and see if we can hoist all the spans in this scope span
-            let span_len = self.scope_spans[0].spans.len();
-            if span_len + count_moved <= split_n {
-                let v = self.scope_spans.remove(0);
-                split_scope_spans.push(v);
-                count_moved += span_len;
-            } else {
-                // We'll need to harvest a subset of the spans
-                let mut spans = vec![];
-                while !self.scope_spans[0].spans.is_empty() && count_moved < split_n {
-                    // Remove the span and add to the new ss
-                    let s = self.scope_spans[0].spans.remove(0);
-                    spans.push(s);
-                    count_moved += 1
-                }
-                // Now we need to clone data from this ScopeSpan for our new ScopeSpans
-                let ss = ScopeSpans {
-                    scope: self.scope_spans[0].scope.clone(),
-                    schema_url: self.scope_spans[0].schema_url.clone(),
-                    spans,
-                };
-                split_scope_spans.push(ss)
-            }
-        }
-        ResourceSpans {
-            scope_spans: split_scope_spans,
-            schema_url: self.schema_url.clone(),
-            resource: self.resource.clone(),
-        }
-    }
-}
-
-impl BatchSizer for ResourceMetrics {
-    fn size_of(&self) -> usize {
-        self.scope_metrics
-            .iter()
-            .flat_map(|sc| &sc.metrics)
-            .map(|m| match &m.data {
-                None => 0,
-                Some(d) => match d {
-                    Data::Gauge(g) => g.data_points.len(),
-                    Data::Sum(s) => s.data_points.len(),
-                    Data::Histogram(h) => h.data_points.len(),
-                    Data::ExponentialHistogram(e) => e.data_points.len(),
-                    Data::Summary(s) => s.data_points.len(),
-                },
-            })
-            .sum()
-    }
-}
-
-impl BatchSizer for [ResourceMetrics] {
-    fn size_of(&self) -> usize {
-        self.iter().map(|n| n.size_of()).sum()
-    }
-}
-
-impl BatchSplittable for ResourceMetrics {
-    fn split(&mut self, split_n: usize) -> Self
-    where
-        Self: Sized,
-    {
-        let mut count_moved = 0;
-        let mut split_scope_metrics = vec![];
-        while !self.scope_metrics.is_empty() && count_moved < split_n {
-            // first just check and see if we can hoist all the spans in this scope span
-            let metric_len = self.scope_metrics[0].metrics.len();
-            if metric_len + count_moved <= split_n {
-                let v = self.scope_metrics.remove(0);
-                split_scope_metrics.push(v);
-                count_moved += metric_len;
-            } else {
-                // We'll need to harvest a subset of the spans
-                let mut metrics = vec![];
-                while !self.scope_metrics[0].metrics.is_empty() && count_moved < split_n {
-                    // Remove the span and add to the new ss
-                    let s = self.scope_metrics[0].metrics.remove(0);
-                    metrics.push(s);
-                    count_moved += 1
-                }
-                // Now we need to clone data from this ScopeSpan for our new ScopeSpans
-                let sm = ScopeMetrics {
-                    scope: self.scope_metrics[0].scope.clone(),
-                    schema_url: self.scope_metrics[0].schema_url.clone(),
-                    metrics,
-                };
-                split_scope_metrics.push(sm)
-            }
-        }
-        ResourceMetrics {
-            scope_metrics: split_scope_metrics,
-            schema_url: self.schema_url.clone(),
-            resource: self.resource.clone(),
-        }
-    }
-}
-
-impl BatchSizer for ResourceLogs {
-    fn size_of(&self) -> usize {
-        self.scope_logs.iter().map(|sc| sc.log_records.len()).sum()
-    }
-}
-
-impl BatchSizer for [ResourceLogs] {
-    fn size_of(&self) -> usize {
-        self.iter().map(|n| n.size_of()).sum()
-    }
-}
-
-impl BatchSplittable for ResourceLogs {
-    fn split(&mut self, split_n: usize) -> Self
-    where
-        Self: Sized,
-    {
-        let mut count_moved = 0;
-        let mut split_scope_logs = vec![];
-        while !self.scope_logs.is_empty() && count_moved < split_n {
-            // first just check and see if we can hoist all the spans in this scope span
-            let span_len = self.scope_logs[0].log_records.len();
-            if span_len + count_moved <= split_n {
-                let v = self.scope_logs.remove(0);
-                split_scope_logs.push(v);
-                count_moved += span_len;
-            } else {
-                // We'll need to harvest a subset of the spans
-                let mut log_records = vec![];
-                while !self.scope_logs[0].log_records.is_empty() && count_moved < split_n {
-                    // Remove the span and add to the new ss
-                    let s = self.scope_logs[0].log_records.remove(0);
-                    log_records.push(s);
-                    count_moved += 1
-                }
-                // Now we need to clone data from this ScopeSpan for our new ScopeSpans
-                let sl = ScopeLogs {
-                    scope: self.scope_logs[0].scope.clone(),
-                    schema_url: self.scope_logs[0].schema_url.clone(),
-                    log_records,
-                };
-                split_scope_logs.push(sl)
-            }
-        }
-        ResourceLogs {
-            scope_logs: split_scope_logs,
-            schema_url: self.schema_url.clone(),
-            resource: self.resource.clone(),
-        }
     }
 }
