+++ src/topology/mod.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod batch;
+pub mod batch_resources;
 pub mod debug;
 pub mod fanout;
 pub mod flush_control;
